renderervk: C++ compatibility for tr_main

+++ code/renderervk/tr_main.c
@@ -578,10 +578,10 @@ void R_SetupProjection( viewParms_t *dest, float zProj, qboolean computeFrustum 
 			stereoSep = 0;
 	}
 
-	ymax = zProj * tan(dest->fovY * M_PI / 360.0f);
+	ymax = zProj * tan((double)(dest->fovY * M_PI / 360.0f));
 	ymin = -ymax;
 
-	xmax = zProj * tan(dest->fovX * M_PI / 360.0f);
+	xmax = zProj * tan((double)(dest->fovX * M_PI / 360.0f));
 	xmin = -xmax;
 
 	width = xmax - xmin;
